% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllQueries.R
\name{findAncestors}
\alias{findAncestors}
\title{Find ATC ancestors for OMOP concept IDs}
\usage{
findAncestors(conn = NULL, conceptIds)
}
\arguments{
\item{conn}{An established OHDSI database connection; NULL defaults to the public OHDSI server}

\item{conceptIds}{A vector of OMOP concept IDs}
}
\description{
Find ATC ancestors for OMOP concept IDs
}
\examples{
\donttest{
findAncestors(conceptIds = c(703470, 705755, 738156))
}
}

