% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllQueries.R
\name{findCommonAncestors}
\alias{findCommonAncestors}
\title{Find common ancestors for OMOP concept IDs}
\usage{
findCommonAncestors(conn = NULL, conceptIds)
}
\arguments{
\item{conn}{An established OHDSI database connection; NULL defaults to the public OHDSI server}

\item{conceptIds}{A vector of OMOP concept IDs}
}
\description{
Find common ancestors for OMOP concept IDs
}
\examples{
\donttest{
findCommonAncestors(conceptIds = c(703470, 705755, 738156))
}
}

