% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllQueries.R
\name{getConceptInformation}
\alias{getConceptInformation}
\title{Retrieve OMOP concept information by ID}
\usage{
getConceptInformation(conn = NULL, conceptIds)
}
\arguments{
\item{conn}{An established OHDSI database connection; NULL defaults to the public OHDSI server}

\item{conceptIds}{A vector of OMOP concept IDs}
}
\description{
Retrieve OMOP concept information by ID
}
\examples{
\donttest{
getConceptInformation(conceptIds = c(19059796,705755))
}
}

