% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllQueries.R
\name{getPublicOhdsiColumnNames}
\alias{getPublicOhdsiColumnNames}
\title{Get field names from a public OHDSI table}
\usage{
getPublicOhdsiColumnNames(conn = NULL, tableName)
}
\arguments{
\item{conn}{An established OHDSI database connection; NULL defaults to the public OHDSI server}

\item{tableName}{An OMOP CDM table name}
}
\description{
Get field names from a public OHDSI table
}
\examples{
\donttest{
getPublicOhdsiColumnNames(tableName = "concept")
}
}

