% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllQueries.R
\name{getPublicOhdsiTableNames}
\alias{getPublicOhdsiTableNames}
\title{Get table names from public OHDSI database}
\usage{
getPublicOhdsiTableNames(conn = NULL)
}
\description{
Get table names from public OHDSI database
}

