% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllQueries.R
\name{getPublicOhdsiUrl}
\alias{getPublicOhdsiUrl}
\title{Get the public OHDSI WebAPI URL}
\usage{
getPublicOhdsiUrl()
}
\value{
URL to the public OHDSI WebAPI
}
\description{
Get the public OHDSI WebAPI URL
}

