% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllQueries.R
\name{createPublicOhdsiConnection}
\alias{createPublicOhdsiConnection}
\title{Create connection to public OHDSI database}
\usage{
createPublicOhdsiConnection()
}
\value{
An established connection to the public OHDSI server
}
\description{
Create connection to public OHDSI database
}
\examples{
\donttest{
conn <- createPublicOhdsiConnection()
}
}

