% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllQueries.R
\name{findCommonAncestors}
\alias{findCommonAncestors}
\title{Find common ancestors for OMOP concept IDs}
\usage{
findCommonAncestors(conn = NULL, conceptIds)
}
\description{
Find common ancestors for OMOP concept IDs
}
\examples{
findCommonAncestors(conceptIds = c(703470, 705755, 738156))
}

