% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllQueries.R
\name{getPublicOhdsiTableNames}
\alias{getPublicOhdsiTableNames}
\title{Get table names from public OHDSI database}
\usage{
getPublicOhdsiTableNames(conn = NULL)
}
\arguments{
\item{conn}{An established OHDSI database connection; NULL defaults to the public OHDSI server}
}
\description{
Get table names from public OHDSI database
}
\examples{
\donttest{
getPublicOhdsiTableNames()
}
}

